refactors arrow udf stuff to make shim easier
