+++ jenkins/databricks/build.sh
@@ -19,10 +19,18 @@ set -e
 
 SPARKSRCTGZ=$1
 # version of Apache Spark we are building against
-BASE_SPARK_VERSION=3.1.1
-BUILD_PROFILES=databricks310
-BUILD_PROFILES=${BUILD_PROFILES:-'databricks301,!snapshot-shims'}
-BASE_SPARK_VERSION=${BASE_SPARK_VERSION:-'3.0.1'}
+BASE_SPARK_VERSION=$2
+BUILD_PROFILES=$3
+BASE_SPARK_VERSION_TO_INSTALL_DATABRICKS_JARS=$4
+BUILD_PROFILES=${BUILD_PROFILES:-'databricks310,!snapshot-shims'}
+BASE_SPARK_VERSION=${BASE_SPARK_VERSION:-'3.1.1'}
+# the version of spark used when we install the databricks jars in .m2
+# 3.1.0-databricks is add because its actually based on Spark 3.1.1
+BASE_SPARK_VERSION_TO_INSTALL_DATABRICKS_JARS=${SPARK_VERSION_TO_INSTALL_DATABRICKS_JARS:-$BASE_SPARK_VERSION}
+SPARK_VERSION_TO_INSTALL_DATABRICKS_JARS=$BASE_SPARK_VERSION_TO_INSTALL_DATABRICKS_JARS-databricks
+# something like spark_3_1 or spark_3_0
+SPARK_MAJOR_VERSION_NUM_STRING=$(echo ${BASE_SPARK_VERSION} | sed 's/\./\_/g' | cut -d _ -f 1,2)
+SPARK_MAJOR_VERSION_STRING=spark_${SPARK_MAJOR_VERSION_NUM_STRING}
 
 echo "tgz is $SPARKSRCTGZ"
 echo "Base Spark version is $BASE_SPARK_VERSION"
@@ -48,8 +56,6 @@ CUDF_VERSION=`mvn help:evaluate -q -pl dist -Dexpression=cudf.version -DforceStd
 SCALA_VERSION=`mvn help:evaluate -q -pl dist -Dexpression=scala.binary.version -DforceStdout`
 CUDA_VERSION=`mvn help:evaluate -q -pl dist -Dexpression=cuda.version -DforceStdout`
 
-# the version of spark used when we install the databricks jars in .m2
-SPARK_VERSION_TO_INSTALL_DATABRICKS_JARS=3.1.0-databricks
 RAPIDS_BUILT_JAR=rapids-4-spark_$SCALA_VERSION-$SPARK_PLUGIN_JAR_VERSION.jar
 RAPIDS_UDF_JAR=rapids-4-spark-udf-examples_$SCALA_VERSION-$SPARK_PLUGIN_JAR_VERSION.jar
 
@@ -61,10 +67,11 @@ CUDF_JAR=${M2DIR}/ai/rapids/cudf/${CUDF_VERSION}/cudf-${CUDF_VERSION}-${CUDA_VER
 
 # pull normal Spark artifacts and ignore errors then install databricks jars, then build again
 JARDIR=/databricks/jars
-SQLJAR=----workspace_spark_3_1--sql--core--core-hive-2.3__hadoop-2.7_${SCALA_VERSION}_deploy.jar
-CATALYSTJAR=----workspace_spark_3_1--sql--catalyst--catalyst-hive-2.3__hadoop-2.7_${SCALA_VERSION}_deploy.jar
-ANNOTJAR=----workspace_spark_3_1--common--tags--tags-hive-2.3__hadoop-2.7_${SCALA_VERSION}_deploy.jar
-COREJAR=----workspace_spark_3_1--core--core-hive-2.3__hadoop-2.7_${SCALA_VERSION}_deploy.jar
+# install the Spark pom file so we get dependencies
+SQLJAR=----workspace_${SPARK_MAJOR_VERSION_STRING}--sql--core--core-hive-2.3__hadoop-2.7_${SCALA_VERSION}_deploy.jar
+CATALYSTJAR=----workspace_${SPARK_MAJOR_VERSION_STRING}--sql--catalyst--catalyst-hive-2.3__hadoop-2.7_${SCALA_VERSION}_deploy.jar
+ANNOTJAR=----workspace_${SPARK_MAJOR_VERSION_STRING}--common--tags--tags-hive-2.3__hadoop-2.7_${SCALA_VERSION}_deploy.jar
+COREJAR=----workspace_${SPARK_MAJOR_VERSION_STRING}--core--core-hive-2.3__hadoop-2.7_${SCALA_VERSION}_deploy.jar
 COREPOM=spark-core_${SCALA_VERSION}-${BASE_SPARK_VERSION}.pom
 COREPOMPATH=$M2DIR/org/apache/spark/spark-core_${SCALA_VERSION}/${BASE_SPARK_VERSION}
 mvn -B install:install-file \
